Limit how big each putRecords request is (#32)
closes #30 

Currently, the strategy is to only respect `CollectionMaxCount` and retry any errors that pop up. If we get more proactive about generating errors maybe we can get more events in.

In practice, we're probably not running into the 5MB for request limit, but the 1MB/second/shard limit. This is more for feature parity with KPL.

* helps prevent quota exceeded errors/ProvisionedThroughputExceededException
* renames some internal variables to be consistent with what the KPL names things. Options will probably be renamed soon
* I'm not sure how the size is counted. Sometimes the partition key length also counts, but what about any other serialization overhead?